requirements.txt
matplotlib~=3.7.0
numpy~=1.24.2
networkx~=3.0
pandas~=1.5.3
biopython
scipy~=1.10.1
scikit-learn~=1.2.1
lmfit~=1.1.0
panel~=0.14.3
build
twine
setuptools~=60.2.0
PyYAML~=6.0
Cerberus~=1.3.4
pandera~=0.13.4
pydantic~=1.10.6